pyproject.toml
[project]
name = "managepro"
version = "0.1.0"
description = "Backend Application for ManagePro"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.33",
    "celery[redis]>=5.5.3",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.1",
    "django-ses>=4.4.0",
    "django-storages[s3]>=1.14.6",
    "django>=5.2.6",
    "djangorestframework-camel-case>=1.4.2",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework>=3.16.1",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.10",
    "redis>=5.2.1",
    "uuid>=1.30",
]
